[frontend] change module version to bigint
Module version is typically in %Y%m%d%H%M%S format (e.g.20171023061952),
so we should be able to store such big integers

+++ frontend/coprs_frontend/coprs/models.py
@@ -1208,7 +1208,7 @@ class Module(db.Model, helpers.Serializer):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), nullable=False)
     stream = db.Column(db.String(100), nullable=False)
-    version = db.Column(db.Integer, nullable=False)
+    version = db.Column(db.BigInteger, nullable=False)
     summary = db.Column(db.String(100), nullable=False)
     description = db.Column(db.Text)
     created_on = db.Column(db.Integer, nullable=True)
